+++ src/main/java/TodoChallengers/BE/challenge/dto/response/UserChallengeResponseDto.java
@@ -1,0 +1,4 @@
+package TodoChallengers.BE.challenge.dto.response;
+
+public class UserChallengeResponseDto {
+}
